ipcache: extract ipcache rest api handler from daemon
This commit extracts the IPCache rest api handler from the daemon
(`daemon/cmd`) to package `pkg/ipcache/api`.
