+++ pkg/ipcache/api/ipcache_api_handler_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package api
 
 import (
 	"net"
